+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irElement.kt
@@ -4,10 +4,13 @@ import com.javiersc.kotlin.compiler.extensions.common.toFqName
 import org.jetbrains.kotlin.ir.IrElement
 import org.jetbrains.kotlin.ir.declarations.IrAnnotationContainer
 import org.jetbrains.kotlin.ir.declarations.IrClass
+import org.jetbrains.kotlin.ir.declarations.IrFunction
+import org.jetbrains.kotlin.ir.declarations.IrValueParameter
 import org.jetbrains.kotlin.ir.expressions.IrFunctionAccessExpression
 import org.jetbrains.kotlin.ir.util.dump
 import org.jetbrains.kotlin.ir.util.dumpKotlinLike
 import org.jetbrains.kotlin.ir.util.hasAnnotation
+import org.jetbrains.kotlin.ir.util.primaryConstructor
 import org.jetbrains.kotlin.name.FqName
 
 public fun Iterable<IrElement>.dump(): String = joinToString("\n") { it.dump() }
@@ -24,3 +27,11 @@ public fun IrElement.hasAnnotation(annotation: FqName): Boolean =
         is IrAnnotationContainer -> hasAnnotation(annotation)
         else -> false
     }
+
+public val IrElement.contextReceivers: List<IrValueParameter>
+    get() =
+        when (this) {
+            is IrFunction -> contextReceivers
+            is IrClass -> primaryConstructor?.contextReceivers.orEmpty()
+            else -> TODO()
+        }